+++ core/src/main/scala/com/github/sharpdata/sharpetl/core/api/WfEvalResult.scala
@@ -1,25 +1,27 @@
 package com.github.sharpdata.sharpetl.core.api
 
+import com.github.sharpdata.sharpetl.core.annotation.Annotations.Evolving
 import com.github.sharpdata.sharpetl.core.repository.model.JobLog
 import com.github.sharpdata.sharpetl.core.syntax.{Formatable, Workflow}
 import com.github.sharpdata.sharpetl.core.util.{Failure, Skipped, Try}
 
-final case class WFInterpretingResult(workflow: Workflow, jobLogs: Seq[Try[JobLog]]) extends Formatable {
+@Evolving("1.0.0")
+final case class WfEvalResult(workflow: Workflow, jobLogs: Seq[Try[JobLog]]) extends Formatable {
 
   override def toString: String = formatString.mkString("\n")
 
   def formatString: Seq[String] = {
     if (jobLogs.nonEmpty) {
-      val jobName = jobLogs.head.get.workflowName
+      val workflowName = jobLogs.head.get.workflowName
       jobLogs.groupBy(_.getClass.getSimpleName)
         .map { case (status, seq) =>
           status match {
-            case "Success" => s"""job name: $jobName SUCCESS x ${seq.size}"""
+            case "Success" => s"""workflow name: $workflowName SUCCESS x ${seq.size}"""
             case "Failure" =>
-              s"""job name: $jobName FAILURE x ${seq.size}, job id: ${seq.head.asInstanceOf[Failure[JobLog]].result.jobId}
+              s"""workflow name: $workflowName FAILURE x ${seq.size}, job id: ${seq.head.asInstanceOf[Failure[JobLog]].result.jobId}
                  | error: ${seq.head.asInstanceOf[Failure[JobLog]].e.getMessage}""".stripMargin
             case "Skipped" =>
-              s"""job name: $jobName SKIPPED x ${seq.size}
+              s"""workflow name: $workflowName SKIPPED x ${seq.size}
                  |from data range start ${seq.head.asInstanceOf[Skipped[JobLog]].result.dataRangeStart}""".stripMargin
           }
         }.toSeq
@@ -29,13 +31,13 @@ final case class WFInterpretingResult(workflow: Workflow, jobLogs: Seq[Try[JobLo
   }
 }
 
-object WFInterpretingResult {
-  def checkSuccessOrThrow(results: Seq[WFInterpretingResult]): Unit = {
+object WfEvalResult {
+  def throwFirstException(results: Seq[WfEvalResult]): Unit = {
     results
       .flatMap(_.jobLogs)
-      .find(it => it.isFailure()).foreach {
-      case Failure(_, throwable) => throw throwable
-      case _ => () //not possible here
-    }
+      .foreach {
+        case Failure(_, throwable) => throw throwable
+        case _ => ()
+      }
   }
 }
