+++ core/src/main/scala/com/github/sharpdata/sharpetl/core/util/Try.scala
@@ -1,7 +1,10 @@
 package com.github.sharpdata.sharpetl.core.util
 
+import com.github.sharpdata.sharpetl.core.annotation.Annotations.Private
+
 import scala.util.control.NonFatal
 
+@Private
 sealed trait Try[+T] extends Product with Serializable {
   def getOrElse[U >: T](default: => U): U
 
@@ -14,6 +17,7 @@ sealed trait Try[+T] extends Product with Serializable {
   def isSkipped(): Boolean
 }
 
+@Private
 object Try {
   def apply[T](f: T => T, t: T): Try[T] =
     try Success(f(t)) catch {
@@ -21,6 +25,7 @@ object Try {
     }
 }
 
+@Private
 final case class Success[+T](result: T) extends Try[T] {
   override def getOrElse[U >: T](default: => U): U = result
 
@@ -33,6 +38,7 @@ final case class Success[+T](result: T) extends Try[T] {
   override def get: T = result
 }
 
+@Private
 final case class Failure[+T](result: T, e: Throwable) extends Try[T] {
   override def getOrElse[U >: T](default: => U): U = default
 
@@ -45,6 +51,7 @@ final case class Failure[+T](result: T, e: Throwable) extends Try[T] {
   override def get: T = result
 }
 
+@Private
 final case class Skipped[+T](result: T) extends Try[T] {
   override def getOrElse[U >: T](default: => U): U = default
