+++ spark/src/main/scala/com/github/sharpdata/sharpetl/spark/cli/Command.scala
@@ -2,8 +2,8 @@ package com.github.sharpdata.sharpetl.spark.cli
 
 import com.github.sharpdata.sharpetl.modeling.cli.{GenerateDwdStepCommand, GenerateSqlFiles}
 import com.github.sharpdata.sharpetl.spark.utils.JavaVersionChecker
-import com.github.sharpdata.sharpetl.core.api.WFInterpretingResult.checkSuccessOrThrow
-import com.github.sharpdata.sharpetl.core.api.{LogDrivenInterpreter, WFInterpretingResult}
+import com.github.sharpdata.sharpetl.core.api.WfEvalResult.throwFirstException
+import com.github.sharpdata.sharpetl.core.api.{LogDrivenInterpreter, WfEvalResult}
 import com.github.sharpdata.sharpetl.core.cli.{BatchJobCommand, EncryptionCommand, SingleJobCommand}
 import com.github.sharpdata.sharpetl.core.notification.NotificationUtil
 import com.github.sharpdata.sharpetl.core.quality.QualityCheckRuleConfig.readQualityCheckRules
@@ -25,14 +25,14 @@ class SingleSparkJobCommand extends SingleJobCommand {
     val interpreter = getSparkInterpreter(local, wfName, releaseResource, etlDatabaseType, readQualityCheckRules())
     JavaVersionChecker.checkJavaVersion()
     try {
-      val wfInterpretingResult: WFInterpretingResult = LogDrivenInterpreter(
+      val wfInterpretingResult: WfEvalResult = LogDrivenInterpreter(
         WorkflowReader.readWorkflow(wfName),
         interpreter,
         jobLogAccessor = jobLogAccessor,
         command = this
       ).eval()
       new NotificationUtil(jobLogAccessor).notify(Seq(wfInterpretingResult))
-      checkSuccessOrThrow(Seq(wfInterpretingResult))
+      throwFirstException(Seq(wfInterpretingResult))
     } finally {
       interpreter.close()
     }
@@ -50,7 +50,7 @@ class BatchSparkJobCommand extends BatchJobCommand {
     val etlDatabaseType = JDBCUtil.dbType
     // val logDrivenInterpreters = if (excelOptions != null) getJobsFromExcel(etlDatabaseType) else getInterpretersFromSqlFile(etlDatabaseType)
     val logDrivenInterpreters = getInterpretersFromSqlFile(etlDatabaseType)
-    val batchJobResult: Seq[WFInterpretingResult] =
+    val batchJobResult: Seq[WfEvalResult] =
       try {
         logDrivenInterpreters.map(_.eval())
       } finally {
@@ -68,7 +68,7 @@ class BatchSparkJobCommand extends BatchJobCommand {
          |""".stripMargin)
     new NotificationUtil(jobLogAccessor).notify(batchJobResult)
     if (failedCount > 0) {
-      checkSuccessOrThrow(batchJobResult)
+      throwFirstException(batchJobResult)
     }
   }
