+++ spark/src/main/scala/com/github/sharpdata/sharpetl/spark/utils/Encoders.scala
@@ -1,12 +1,7 @@
 package com.github.sharpdata.sharpetl.spark.utils
 
-import com.github.sharpdata.sharpetl.core.quality.{CheckResult, DataQualityCheckResult}
-import org.apache.spark.sql.{DataFrame, Encoder, Encoders}
-
-final class SparkCheckResult(override val warn: Seq[DataQualityCheckResult],
-                             override val error: Seq[DataQualityCheckResult],
-                             override val passed: DataFrame)
-  extends CheckResult(warn, error, passed)
+import com.github.sharpdata.sharpetl.core.quality.DataQualityCheckResult
+import org.apache.spark.sql.{Encoder, Encoders}
 
 object Encoder extends Serializable {
   val dqEncoder: Encoder[DataQualityCheckResult] = Encoders.product[DataQualityCheckResult]
